+++ src/componenets/preview_strip.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useState, useEffect} from 'react';
 import ImageButton from './image_button';
 
 import Button from '@material-ui/core/Button';
@@ -39,91 +39,99 @@ const styles = theme => ({
       opacity: 1.0,
     }
   },
-
 });
 
+// if there are enough to show: |L|**|R|R|R|
+// else try to show all five
 const showThisMany = 5;
+const idealIndex = 1;
 
-class previewStrip  extends React.Component {
-  constructor(props) {
-    super(props);
+const previewStrip = props => {
+  const {items, id, assets, collection, ensemble, classes} = props;
+  const [indexOffset, setIndexOffset] = useState(0);
 
-    // now playing starts out in pos 2; skip from 1-3
-    this.state = {
-      indexOffset: 2,
-    };
+  useEffect(() => {
+      setIndexOffset(0);
+    },[id],
+  );
+
+  const incrementIndex = () => {
+      setIndexOffset(indexOffset + 1);
   };
 
-  componentDidUpdate(prevProps) {
-    const {id, items} = this.props;
-    if (id && (prevProps.id !== id)) {
-      this.setState({indexOffset: 2});
-      return;
+  const decrementIndex = () => {
+    setIndexOffset(indexOffset - 1);
+  };
+
+  const videoCount = items.length;
+  let currentVideoIndex = 0;
+  for (var i = 0; i < videoCount; i++) {
+    if (items[i] && items[i].id === id) {
+      currentVideoIndex = i;
+      break;
     }
-    // any prop change, reset scroll buttons
-    if (items && prevProps.items && (prevProps.items.length !== items.length)) {
-      this.setState({indexOffset: 2});
+  }
+  currentVideoIndex += indexOffset;
+
+  // layout "now playing" on strip based on availability
+  // how many we got?
+  let numRight = videoCount - currentVideoIndex;
+  let numLeft =  currentVideoIndex;
+
+  let leftIndex = currentVideoIndex;
+  if (numLeft + numRight  < showThisMany) {
+    // if there are not enough to scroll, show all
+    leftIndex = 0;
+  } else {
+    // else fill the strip as best as we can
+    // grab showThisMany - 1 neighbors
+    for (var j = 1; j < showThisMany;) {
+      if (numRight > 0) {
+        // grab one from the right
+        numRight--;
+        j++;
+      }
+      if (numLeft > 0 && j < showThisMany) {
+        // grab one from the left
+        numLeft--;
+        j++;
+        // move it on over
+        leftIndex--;
+      }
     }
   }
 
-  incrementIndex = () => {
-    this.setIndexOffset(this.state.indexOffset + 1);
-  };
+  const rightIndex = Math.min(videoCount, leftIndex + showThisMany);
+  const show = items.slice(leftIndex, rightIndex);
 
-  decrementIndex = () => {
-    this.setIndexOffset(this.state.indexOffset - 1);
-  };
-
-  setIndexOffset = offset => {
-    this.setState({indexOffset: offset});
-  };
-
-  render() {
-    const {items, id, assets, classes} = this.props;
-    const {indexOffset} = this.state;
-
-    const videoCount = items.length;
-    let currentVideoIndex = 0;
-    for (var i = 0; i < videoCount; i++) {
-      if (items[i] && items[i].id === id) {
-        currentVideoIndex = i;
-        break;
-      }
+  const images = show.map(i => {
+    return {
+      id: i.id,
+      url: assets + i.poster,
+      title: i.title,
+      width: '200px'
     }
-    const padSize = Math.floor(showThisMany / 2);
-    const firstIndex = Math.max(0, currentVideoIndex - padSize + indexOffset);
-    const lastIndex = Math.min(videoCount, firstIndex + showThisMany);
-    const show = items.slice(firstIndex, lastIndex);
+  });
 
-    const images = show.map(i => {
-      return {
-        id: i.id,
-        url: assets + i.poster,
-        title: i.title,
-        width: '200px'
-      }
-    });
+  return (
+    <div className={classes.panel}>
+      <Button className={classes.navButton}
+              key="up"
+              onClick={decrementIndex}
+              disabled={leftIndex <= 0}>
+        {(leftIndex > 0) && <UpIcon />}
+      </Button>
+      {images.map(image => {
+        return (<Link key={'link_' + image.id} to={{search: '?collection=' + collection + '&ensemble=' + ensemble + '&id=' + image.id}}>
+          <ImageButton image={image} nowPlaying={image.id === id}/>
+        </Link>)
+      })}
+      {(rightIndex < videoCount) &&
+        <Button className={classes.navButton} key="down" onClick={incrementIndex}>
+          <DownIcon />
+        </Button>}
+    </div>
+  );
+};
 
-    return (
-      <div className={classes.panel}>
-        <Button className={classes.navButton}
-                key="up"
-                onClick={this.decrementIndex}
-                disabled={firstIndex <= 0}>
-          {(firstIndex > 0) && <UpIcon />}
-        </Button>
-        {images.map(image => {
-          return (<Link to={{search: 'id=' + image.id}}>
-            <ImageButton image={image} nowPlaying={image.id === id}/>
-          </Link>)
-        })}
-        {(lastIndex < videoCount) &&
-          <Button className={classes.navButton} key="down" onClick={this.incrementIndex}>
-            <DownIcon />
-          </Button>}
-      </div>
-    );
-  };
-}
 export default withStyles(styles)(previewStrip);
-
